Part 2New/inference_server_project-starter/pyproject.toml
[tool.poetry]
name = "inference-server"
version = "0.1.0"
description = "A simple image classification inference server"
authors = ["Tian Guo (Mac studio) <tian@wpi.edu>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
torch = "^2.5.1"
torchvision = "^0.20.1"
pillow = "^11.0.0"
opencv-python-headless = "^4.10.0.84"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
